Moved update_text_roots to new text_hierarchy module
